refactor(worker_threads): split worker.ts into per-concern modules (873 → 552 + 340)
Same per-concern split pattern. The parent-side `Worker` class is
one cohesive Gio.Subprocess-driven implementation (spawn, postMessage,
terminate, stdin/stdout/stderr readers, SCM_RIGHTS fd shipping,
cross-process MessagePort routing, cleanup) — every method touches
the per-instance subprocess pipes + the SAB socketpair + the port
registry, so it stays together. The clear extraction target is the
~330-LoC `BOOTSTRAP_CODE` JS-as-string template: the entire JS body
that runs in the spawned gjs child lives in its own runtime; the
parent class never reads it, only encodes the string and hands the
temp .mjs path to Gio.SubprocessLauncher.

Layout:

  src/worker.ts            (552 LoC) — Worker class + WorkerOptions
                                       interface + parent-side IPC
                                       (was 873).
  src/worker-bootstrap.ts  (340 LoC) — `export const BOOTSTRAP_CODE`
                                       — the child-process body
                                       (MainLoop + async stdin reader
                                       + materialise() walker +
                                       makeSharedBuffer() /
                                       makeChildSidePort() factories
                                       + SCM_RIGHTS fd-drain logic +
                                       parentPort proxy).

`BOOTSTRAP_BYTES = _encoder.encode(BOOTSTRAP_CODE)` stays in worker.ts
next to its sole consumer (the temp-file write in the Worker
constructor). The original comment block explaining the side-channel
ordering and the lazy `imports.gi.GjsifySabNative` trick moves
verbatim into the bootstrap module's header.

Public surface preserved verbatim; no exports changed.

286 GJS + 281 Node worker_threads tests still green; tsc + build clean.

STATUS.md: new Completed entry.
