+++ STATUS.md
@@ -405,6 +405,8 @@ Not yet implemented (but potentially relevant for GJS projects):
 
 ### Completed
 
+- ~~**`@gjsify/worker_threads` — split `src/worker.ts` into per-concern modules (873 → 552 + 340)**~~✓ — Same per-concern split pattern. Base `worker.ts` shrinks from 873 LoC to 552 LoC; the parent-side `Worker` class + WorkerOptions interface + IPC plumbing (spawn, postMessage, terminate, stdin/stdout/stderr readers, `_extractCrossProcessPorts`, `_serializeWithSabTransfer`, `_resolveFilename`, `_cleanup`) stay together because they share tight coupling with the per-instance `Gio.Subprocess`, the `_sabSocketFd`, and the `_portRegistry: CrossProcessPortRegistry`. The clear extraction target is the ~330-LoC `BOOTSTRAP_CODE` JS-as-string template — the entire JS body that runs in the spawned gjs child (`import GLib from 'gi://GLib'`, the `MainLoop`, async stdin reader, `materialise()` walker, `makeSharedBuffer()` / `makeChildSidePort()` / `parentPort` factories, the SCM_RIGHTS fd-drain logic) lives in its own runtime; the parent class never reads it, only encodes the string and hands the path to `Gio.SubprocessLauncher.spawnv`. New module `packages/node/worker_threads/src/worker-bootstrap.ts` (340 LoC including its file header + the verbatim preserved comment block explaining the side-channel ordering and the lazy `imports.gi.GjsifySabNative` trick) exports a single `BOOTSTRAP_CODE` const. The base file imports it via `import { BOOTSTRAP_CODE } from './worker-bootstrap.js';` — `BOOTSTRAP_BYTES = _encoder.encode(BOOTSTRAP_CODE)` stays in `worker.ts` next to its sole consumer (the temp-file write in the `Worker` constructor). Public surface preserved verbatim; no exports changed. 286 GJS + 281 Node worker_threads tests still green; tsc + build clean. Same approach as the `@gjsify/buffer` `buffer-codec.ts` extraction (PR #283) — pull out a self-contained subprocess-side concern that has zero state-coupling to the parent class, so the base file becomes easier to reason about without losing any expressive power.
+
 - ~~**`@gjsify/buffer` — extract `buffer-codec.ts` from `buffer.ts` (817 → 701 + 140)**~~✓ — Same per-concern split pattern (each module owns a clear slice). The Buffer class proper is one cohesive `Uint8Array` subclass implementation, so the win comes from extracting the two encoding-codec functions that aren't class methods. Layout: `buffer-codec.ts` (140 LoC) holds `encodeString(str, encoding)` and `decodeString(buf, encoding, start?, end?)` — the per-encoding String↔Uint8Array switch logic for utf8/ascii/latin1/base64/base64url/hex/utf16le — plus the small `checkOffset(offset, ext, length)` range-check helper. `buffer.ts` shrinks from 817 LoC to 701 LoC; its imports section is noticeably less cluttered (the `textEncoder` / `textDecoder` module-locals live in `buffer-codec.ts` now; the 2 remaining call sites in `Buffer.byteLength` use a one-shot `new TextEncoder()` — TextEncoder is cheap to construct, and `byteLength` is rarely on a hot path). Public surface preserved verbatim; no exports changed. 326 GJS buffer tests still green; tsc + build clean.
 
 - ~~**`README.md` modernized — `npm`/`yarn` commands replaced with `gjsify <subcommand>`**~~✓ — Top-level user-facing README brought in line with v0.4.21's Node-free CLI. Quick Start now leads with the `curl … install.mjs | gjs -m -` Node-free bootstrap; the legacy `npm install -g @gjsify/cli` is demoted to a clearly-marked escape hatch. New-project flow is `gjsify create` + `gjsify install --immutable` + `gjsify build` + `gjsify run start` (was `npm create @gjsify/app` + `npm install` + `npm run build` + `npm start`). Prerequisites section drops `nodejs` from the Fedora/Ubuntu apt/dnf one-liners and labels Node.js itself as optional (only needed for the cross-validation `test:node` track or for managing `@gjsify/cli` via `npm install -g`). Development section: `yarn build` → `gjsify foreach -A -t build`, `yarn check` → `gjsify check`, `yarn test` → `gjsify foreach -A test`, `yarn test:integration*` → per-suite `gjsify workspace <name> test`, per-package `yarn test:{node,gjs}` → `gjsify test [--runtime gjs|node]` with a paragraph explaining the new `gjsify test` runner. Verified the cited commands by running each of `gjsify --help` / `install --help` / `dlx --help` / `run --help` / `foreach --help` / `workspace --help` / `check --help` / `test --help` / `build --help` / `create --help` against the committed bundle.
